+++ packages/hooks/src/useAddress.ts
@@ -21,13 +21,17 @@ export function getBlockExplorerAddressLink(network: Chain, address: string) {
 
 // make the chain optional, if not provided, it will use from wagmi conig
 export const useAddress = (UseAddressOptions: UseAddressOptions) => {
-  const checkSumAddress = UseAddressOptions?.address ? getAddress(UseAddressOptions.address) : undefined;
+  const checkSumAddress =
+    UseAddressOptions?.address &&
+    (isAddress(UseAddressOptions?.address) ? getAddress(UseAddressOptions.address) : undefined);
+
+  const isValidAddress = Boolean(checkSumAddress);
 
   const { data: ens, isLoading: isEnsNameLoading } = useEnsName({
     address: checkSumAddress,
     chainId: 1,
     query: {
-      enabled: isAddress(checkSumAddress ?? ""),
+      enabled: isValidAddress,
     },
   });
 
@@ -42,11 +46,11 @@ export const useAddress = (UseAddressOptions: UseAddressOptions) => {
 
   const shortAddress = checkSumAddress ? `${checkSumAddress.slice(0, 6)}...${checkSumAddress.slice(-4)}` : undefined;
 
-  const isValidAddress = checkSumAddress ? isAddress(checkSumAddress) : false;
-  const blockExplorerAddressLink = getBlockExplorerAddressLink(
-    UseAddressOptions?.chain ?? mainnet,
-    checkSumAddress ?? "",
-  );
+  const blockExplorerAddressLink = checkSumAddress
+    ? getBlockExplorerAddressLink(UseAddressOptions?.chain ?? mainnet, checkSumAddress)
+    : "";
+
+  const blockieUrl = checkSumAddress ? blo(checkSumAddress) : undefined;
 
   return {
     checkSumAddress,
@@ -56,6 +60,6 @@ export const useAddress = (UseAddressOptions: UseAddressOptions) => {
     blockExplorerAddressLink,
     isValidAddress,
     shortAddress,
-    blockieUrl: UseAddressOptions.address ? blo(UseAddressOptions.address as `0x${string}`) : undefined,
+    blockieUrl,
   };
 };