+++ packages/debug-contracts/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @scaffold-ui/debug-contracts
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [5e23fdf]
+  - @scaffold-ui/hooks@0.1.4
+  - @scaffold-ui/components@0.1.5
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [c805b7e]
+  - @scaffold-ui/hooks@0.1.3
+  - @scaffold-ui/components@0.1.4
+
 ## 0.1.2
 
 ### Patch Changes